[crypto] Extract X509 functions for mbedTLS and PSA
PSA crypto backend uses mbedTLS API to implement functions
for certificate generation and parsing, so the functions
are duplicated between PSA and mbedTLS backends.

Extract the functions to reduce the code duplication.
