+++ src/crypto/BUILD.gn
@@ -100,7 +100,11 @@ if (chip_crypto == "openssl") {
   import("//build_overrides/mbedtls.gni")
 
   source_set("cryptopal_mbedtls") {
-    sources = [ "CHIPCryptoPALmbedTLS.cpp" ]
+    sources = [
+      "CHIPCryptoPALmbedTLS.cpp",
+      "CHIPCryptoPALmbedTLSCert.cpp",
+      "CHIPCryptoPALmbedTLS.h",
+    ]
     public_deps = [ ":public_headers" ]
 
     if (!chip_external_mbedtls) {
@@ -112,10 +116,10 @@ if (chip_crypto == "openssl") {
 
   source_set("cryptopal_psa") {
     sources = [
+      "CHIPCryptoPALmbedTLSCert.cpp",
+      "CHIPCryptoPALmbedTLS.h",
       "CHIPCryptoPALPSA.cpp",
       "CHIPCryptoPALPSA.h",
-      "PSAOperationalKeystore.cpp",
-      "PSAOperationalKeystore.h",
     ]
     public_deps = [ ":public_headers" ]
 
@@ -139,6 +143,8 @@ static_library("crypto") {
 
   if (chip_crypto == "psa") {
     sources += [
+      "PSAOperationalKeystore.cpp",
+      "PSAOperationalKeystore.h",
       "PSASessionKeystore.cpp",
       "PSASessionKeystore.h",
     ]
